--- mock/dashboard/workplace/queryActivities.js
@@ -1,118 +1,0 @@
-const data = {
-  data: [
-    {
-      id: 'trend-1',
-      updatedAt: '2024-11-27T10:11:47.672Z',
-      user: {
-        name: '曲丽丽',
-        avatar:
-          'https://gw.alipayobjects.com/zos/rmsportal/BiazfanxmamNRoxxVxka.png',
-      },
-      group: {
-        name: '高逼格设计天团',
-        link: 'http://github.com/',
-      },
-      project: {
-        name: '六月迭代',
-        link: 'http://github.com/',
-      },
-      template: '在 @{group} 新建项目 @{project}',
-    },
-    {
-      id: 'trend-2',
-      updatedAt: '2024-11-27T10:11:47.672Z',
-      user: {
-        name: '付小小',
-        avatar:
-          'https://gw.alipayobjects.com/zos/rmsportal/cnrhVkzwxjPwAaCfPbdc.png',
-      },
-      group: {
-        name: '高逼格设计天团',
-        link: 'http://github.com/',
-      },
-      project: {
-        name: '六月迭代',
-        link: 'http://github.com/',
-      },
-      template: '在 @{group} 新建项目 @{project}',
-    },
-    {
-      id: 'trend-3',
-      updatedAt: '2024-11-27T10:11:47.672Z',
-      user: {
-        name: '林东东',
-        avatar:
-          'https://gw.alipayobjects.com/zos/rmsportal/gaOngJwsRYRaVAuXXcmB.png',
-      },
-      group: {
-        name: '中二少女团',
-        link: 'http://github.com/',
-      },
-      project: {
-        name: '六月迭代',
-        link: 'http://github.com/',
-      },
-      template: '在 @{group} 新建项目 @{project}',
-    },
-    {
-      id: 'trend-4',
-      updatedAt: '2024-11-27T10:11:47.672Z',
-      user: {
-        name: '周星星',
-        avatar:
-          'https://gw.alipayobjects.com/zos/rmsportal/WhxKECPNujWoWEFNdnJE.png',
-      },
-      project: {
-        name: '5 月日常迭代',
-        link: 'http://github.com/',
-      },
-      template: '将 @{project} 更新至已发布状态',
-    },
-    {
-      id: 'trend-5',
-      updatedAt: '2024-11-27T10:11:47.672Z',
-      user: {
-        name: '朱偏右',
-        avatar:
-          'https://gw.alipayobjects.com/zos/rmsportal/ubnKSIfAJTxIgXOKlciN.png',
-      },
-      project: {
-        name: '工程效能',
-        link: 'http://github.com/',
-      },
-      comment: {
-        name: '留言',
-        link: 'http://github.com/',
-      },
-      template: '在 @{project} 发布了 @{comment}',
-    },
-    {
-      id: 'trend-6',
-      updatedAt: '2024-11-27T10:11:47.672Z',
-      user: {
-        name: '乐哥',
-        avatar:
-          'https://gw.alipayobjects.com/zos/rmsportal/jZUIxmJycoymBprLOUbT.png',
-      },
-      group: {
-        name: '程序员日常',
-        link: 'http://github.com/',
-      },
-      project: {
-        name: '品牌迭代',
-        link: 'http://github.com/',
-      },
-      template: '在 @{group} 新建项目 @{project}',
-    },
-  ],
-};
-module.exports = {
-  'GET /api/activities': (req, res) => {
-    setTimeout(
-      () => {
-        res.json(data);
-      },
-      Math.random(5, 10) * 1000,
-    );
-  },
-};
