--- src/services/account/settings.ts
@@ -1,20 +1,0 @@
-import type { CurrentUser, GeographicItemType } from '@/types/account/settings';
-import { request } from '@/utils/request';
-
-export async function queryCurrent(): Promise<{ data: CurrentUser }> {
-  return request('/api/accountSettingCurrentUser');
-}
-
-export async function queryProvince(): Promise<{ data: GeographicItemType[] }> {
-  return request('/api/geographic/province');
-}
-
-export async function queryCity(
-  province: string,
-): Promise<{ data: GeographicItemType[] }> {
-  return request(`/api/geographic/city/${province}`);
-}
-
-export async function query() {
-  return request('/api/users');
-}
