--- src/services/dashboard/analysis.ts
@@ -1,6 +1,0 @@
-import type { AnalysisData } from '@/types/dashboard/analysis';
-import { request } from '@/utils/request';
-
-export async function fakeChartData(): Promise<{ data: AnalysisData }> {
-  return request('/api/fake_analysis_chart_data');
-}
