--- src/services/dashboard/workplace.ts
@@ -1,14 +1,0 @@
-import { request } from '@/utils/request';
-import type { ActivitiesType, AnalysisData, NoticeType } from './data';
-
-export async function queryProjectNotice(): Promise<{ data: NoticeType[] }> {
-  return request('/api/project/notice');
-}
-
-export async function queryActivities(): Promise<{ data: ActivitiesType[] }> {
-  return request('/api/activities');
-}
-
-export async function fakeChartData(): Promise<{ data: AnalysisData }> {
-  return request('/api/fake_workplace_chart_data');
-}
