--- src/services/form/advanced-form.ts
@@ -1,8 +1,0 @@
-import { request } from '@/utils/request';
-
-export async function fakeSubmitForm(params: any) {
-  return request('/api/advancedForm', {
-    method: 'POST',
-    data: params,
-  });
-}
