--- src/services/form/basic-form.ts
@@ -1,8 +1,0 @@
-import { request } from '@/utils/request';
-
-export async function fakeSubmitForm(params: any) {
-  return request('/api/basicForm', {
-    method: 'POST',
-    data: params,
-  });
-}
