+++ src/services/login.ts
@@ -1,16 +1,9 @@
-import { request } from '@/utils/request';
-
 export async function getFakeCaptcha(
   params: {
     phone?: string;
   },
   options?: { [key: string]: any },
 ) {
-  return request<API.FakeCaptcha>('/api/login/captcha', {
-    method: 'GET',
-    params: {
-      ...params,
-    },
-    ...(options || {}),
-  });
+  const newparams = new URLSearchParams(params).toString();
+  return fetch(`/api/login/captcha?${newparams}`);
 }
