--- src/types/dashboard/monitor.ts
@@ -1,5 +1,0 @@
-export type TagType = {
-  name: string;
-  value: number;
-  type: string;
-};
