--- src/types/form/step-form.ts
@@ -1,9 +1,0 @@
-export interface StepDataType {
-  payAccount: string;
-  receiverAccount: string;
-  receiverName: string;
-  amount: string;
-  receiverMode: string;
-}
-
-export type CurrentTypes = 'base' | 'confirm' | 'result';
