+++ src/types/index.ts
@@ -1,3 +1,5 @@
+import { DataItem } from '@antv/g2plot/esm/interface/config';
+
 export type BasicGood = {
   id: string;
   name?: string;
@@ -130,3 +132,140 @@ export type CardListItemDataType = {
   content: string;
   members: Member[];
 };
+
+export type tabKeyType = 'articles' | 'applications' | 'projects';
+export interface TagType {
+  key: string;
+  label: string;
+}
+
+export type NoticeType = {
+  id: string;
+  title: string;
+  logo: string;
+  description: string;
+  updatedAt: string;
+  member: string;
+  href: string;
+  memberLink: string;
+};
+
+export type CurrentUser = {
+  name: string;
+  avatar: string;
+  userid: string;
+  notice: NoticeType[];
+  email: string;
+  signature: string;
+  title: string;
+  group: string;
+  tags: TagType[];
+  notifyCount: number;
+  unreadCount: number;
+  country: string;
+  geographic: GeographicType;
+  address: string;
+  phone: string;
+};
+
+export type GeographicItemType = {
+  name: string;
+  id: string;
+};
+
+export type GeographicType = {
+  province: GeographicItemType;
+  city: GeographicItemType;
+};
+
+export interface VisitDataType {
+  x: string;
+  y: number;
+}
+
+export type SearchDataType = {
+  index: number;
+  keyword: string;
+  count: number;
+  range: number;
+  status: number;
+};
+
+export interface OfflineChartData {
+  date: number;
+  type: number;
+  value: number;
+}
+
+export type RadarData = {
+  name: string;
+  label: string;
+  value: number;
+};
+
+export interface AnalysisData {
+  visitData: DataItem[];
+  visitData2: DataItem[];
+  salesData: DataItem[];
+  searchData: DataItem[];
+  offlineData: OfflineDataType[];
+  offlineChartData: DataItem[];
+  salesTypeData: DataItem[];
+  salesTypeDataOnline: DataItem[];
+  salesTypeDataOffline: DataItem[];
+  radarData: RadarData[];
+}
+
+export type MonitorTagType = {
+  name: string;
+  value: number;
+  type: string;
+};
+
+export { DataItem };
+
+export interface TagType {
+  key: string;
+  label: string;
+}
+
+export type OfflineDataType = {
+  name: string;
+  cvr: number;
+};
+
+export type ActivitiesType = {
+  id: string;
+  updatedAt: string;
+  user: {
+    link?: string;
+    name: string;
+    avatar: string;
+  };
+  group: {
+    name: string;
+    link: string;
+  };
+  project: {
+    name: string;
+    link: string;
+  };
+
+  template: string;
+};
+
+export type RadarDataType = {
+  label: string;
+  name: string;
+  value: number;
+};
+
+export interface StepDataType {
+  payAccount: string;
+  receiverAccount: string;
+  receiverName: string;
+  amount: string;
+  receiverMode: string;
+}
+
+export type CurrentTypes = 'base' | 'confirm' | 'result';
